Update recipes.json to replace acetylene with a dissecting microscope in the Labs Catalyst
Loose acetylene gas can't be used as a crafting ingredient, and the game isn't currently set up to use charges from tanks as an ingredient in a crafting recipe. I've replaced it with a dissecting microscope. It's perhaps a little bit harder to find the first one, as they're mostly found in environments crowded with enemies, but they're commonly found in labs so you can reasonably chain runs once you're on a roll.